Move S3 client to new location
