+++ src/bos/operators/utils/boot_image_metadata/s3_boot_image_metadata.py
@@ -25,10 +25,10 @@ import logging
 
 from botocore.exceptions import ClientError
 
+from bos.common.clients.s3 import S3BootArtifacts, S3MissingConfiguration, S3Url, \
+                                           ArtifactNotFound
 from bos.common.utils import exc_type_msg
 from bos.operators.utils.boot_image_metadata import BootImageMetaData, BootImageMetaDataBadRead
-from bos.operators.utils.clients.s3 import S3BootArtifacts, S3MissingConfiguration, S3Url, \
-                                           ArtifactNotFound
 
 LOGGER = logging.getLogger(__name__)
