+++ src/bos/server/controllers/v2/boot_set/artifacts.py
@@ -22,9 +22,9 @@
 # OTHER DEALINGS IN THE SOFTWARE.
 #
 
+from bos.common.clients.s3 import S3Object, ArtifactNotFound
 from bos.common.utils import exc_type_msg
 from bos.operators.utils.boot_image_metadata.factory import BootImageMetaDataFactory
-from bos.operators.utils.clients.s3 import S3Object, ArtifactNotFound
 
 from .defs import LOGGER
 from .exceptions import BootSetError, BootSetWarning
